+++ src/server/game/Reputation/ReputationMgr.cpp
@@ -285,7 +285,7 @@ bool ReputationMgr::SetReputation(FactionEntry const* factionEntry, int32 standi
     // if spillover definition exists in DB, override DBC
     if (const RepSpilloverTemplate* repTemplate = sObjectMgr->GetRepSpilloverTemplate(factionEntry->ID))
     {
-        for (uint32 i = 0; i < MAX_SPILLOVER_FACTIONS; ++i)
+        for (uint8 i = 0; i < MAX_SPILLOVER_FACTIONS; ++i)
         {
             if (repTemplate->faction[i])
             {
